+++ main.py
@@ -1,2 +1,4 @@
 from tennis.wta.parser import ParsePage
+from tennis.wta.models import Tournament
 
+print(Tournament(1, 'test', 'October 16, 2019', 'R', 'H', 1, 1))