+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### Version 1.1
 
 * Refactoring of the different modules. Added _wta.tennis.models_.
+* Created a _check\_path()_ function for the output folder
 
 ### Version 1